+++ lib/src/fields/unified_async_picker_field.dart
@@ -365,6 +365,7 @@ class _UnifiedAsyncPickerFieldState<T>
         placeholder: widget.placeholder ?? dec.placeholder ?? widget.label,
         labelStyle: dec.labelStyle,
         style: dec.fieldStyle,
+        placeholderStyle: dec.placeholderStyle,
         backgroundColor: dec.backgroundColor,
         headerBackgroundColor:
           dec.headerBackgroundColor ?? dec.backgroundColor,
@@ -752,6 +753,7 @@ class _UnifiedAsyncMultiPickerFieldState<T>
         placeholder: widget.placeholder ?? dec.placeholder ?? widget.label,
         labelStyle: dec.labelStyle,
         style: dec.fieldStyle,
+        placeholderStyle: dec.placeholderStyle,
         backgroundColor: dec.backgroundColor,
         headerBackgroundColor:
           dec.headerBackgroundColor ?? dec.backgroundColor,