+++ lib/src/fields/unified_customizable_async_picker_field.dart
@@ -353,6 +353,7 @@ class _UnifiedCustomizableAsyncPickerFieldState<T>
         placeholder: widget.placeholder ?? dec.placeholder ?? widget.label,
         labelStyle: dec.labelStyle,
         style: dec.fieldStyle,
+        placeholderStyle: dec.placeholderStyle,
         backgroundColor: dec.backgroundColor,
         headerBackgroundColor:
           dec.headerBackgroundColor ?? dec.backgroundColor,
@@ -723,6 +724,7 @@ class _UnifiedCustomizableAsyncMultiPickerFieldState<T>
         placeholder: widget.placeholder ?? dec.placeholder ?? widget.label,
         labelStyle: dec.labelStyle,
         style: dec.fieldStyle,
+        placeholderStyle: dec.placeholderStyle,
         backgroundColor: dec.backgroundColor,
         headerBackgroundColor:
           dec.headerBackgroundColor ?? dec.backgroundColor,