+++ lib/src/fields/unified_time_of_day_field.dart
@@ -300,6 +300,7 @@ class _UnifiedTimeOfDayFieldState extends State<UnifiedTimeOfDayField> {
           placeholder: widget.placeholder ?? d.placeholder ?? d.label,
           labelStyle: d.labelStyle,
           style: d.fieldStyle,
+          placeholderStyle: d.placeholderStyle,
           backgroundColor: d.backgroundColor,
           headerBackgroundColor:
           d.headerBackgroundColor ?? d.backgroundColor,